refactor: extract evaluator_helpers.py from evaluator.py
Move standalone scoring functions (_count_contradictions,
_check_first_turn_character, _sample_narratives, _compute_composite,
_compute_verdict) and CATEGORY_LABELS to evaluator_helpers.py (170 lines).

evaluator.py: 611 → 478 lines (-133, -21.8%).
